Consolidate /education into /learn; fix orphaned articles/blog content
- Merge app/education into app/learn: move 8 why-* pages and
  pathway-hub.tsx, port evidence-literacy content to a static page,
  and replace app/learn/page.tsx (previously self-contradictory,
  canonical pointed at /guides/) with the corrected neuroscience hub
  plus a new "Recent stack guides" section surfacing 5 previously
  orphaned learn posts.
- Delete dead education infrastructure (lib/education.ts,
  app/education/**, content/education/**) now that all unique content
  lives under /learn.
- Add redirects from every old /education/* path to its /learn
  equivalent, plus a catch-all fallback.
- Extend content-collections.ts to include sibling .md files (not just
  .mdx) for articles, and add a new blogPosts collection for
  content/blog — this was the root cause of ~37 sitemap entries
  (data/articles/articles.json, data/blog/posts.json) having no
  corresponding rendered route. Sanitize literal "<" in markdown so it
  doesn't break MDX compilation, and estimate reading time when absent.
- Point build-search-index.mjs's education indexer at app/learn
  instead of the now-deleted app/education, and exclude dynamic
  [slug] routes from the generated docs.
