--- app/education/layout.tsx
@@ -1,5 +1,0 @@
-import type { ReactNode } from 'react'
-
-export default function EducationLayout({ children }: { children: ReactNode }) {
-  return children
-}
