--- content/education/evidence-levels.md
@@ -1,36 +1,0 @@
----
-slug: evidence-levels
-title: Understanding Evidence Levels
-summary: How to read the evidence grades used across this site — what separates strong human trial data from preliminary mechanistic or traditional-use claims, and why conservative labels matter.
-category: Research Methodology
-goals: []
-pathways: []
-evidenceGrade: Educational
-safety: Educational overview only — it does not endorse any supplement.
-keywords:
-  - evidence grade
-  - clinical trial
-  - meta-analysis
-  - mechanism
-  - confidence
-  - study quality
-tags:
-  - methodology
-  - evidence
-relatedHerbs: []
-relatedCompounds: []
-readingTime: 6 min read
-updated: 2026-06-11
----
-
-Not all evidence is equal. This site grades claims so you can tell the difference
-between a well-replicated human finding and a hopeful extrapolation.
-
-- **Strong** — multiple randomized controlled trials or meta-analyses in humans.
-- **Moderate** — a few human trials, sometimes small or with mixed results.
-- **Limited** — early human data, observational signals, or single studies.
-- **Preliminary** — mostly mechanism, animal, or traditional-use evidence.
-
-When source data is incomplete, labels stay conservative on purpose. A low grade
-does not mean a supplement "does not work" — it means the human evidence is not
-yet strong enough to claim that it does.
