--- content/education/how-neurotransmitters-work.md
@@ -1,50 +1,0 @@
----
-slug: how-neurotransmitters-work
-title: How Neurotransmitters Work
-summary: A plain-language explanation of how neurotransmitters carry signals between neurons, why receptor balance matters more than any single chemical, and how this framing applies to herbs and compounds.
-category: Foundational Neuroscience
-goals:
-  - focus
-  - mood
-  - stress
-pathways:
-  - dopamine
-  - serotonin
-  - gaba
-  - glutamate
-evidenceGrade: Educational
-safety: Educational overview only — not dosing or treatment guidance.
-keywords:
-  - neurotransmitter
-  - synapse
-  - receptor
-  - reuptake
-  - signaling
-  - neurochemistry basics
-tags:
-  - neuroscience
-  - basics
-relatedHerbs:
-  - ashwagandha
-  - bacopa
-relatedCompounds:
-  - l-theanine
-  - magnesium-glycinate
-readingTime: 7 min read
-updated: 2026-06-10
----
-
-Neurotransmitters are chemical messengers that neurons use to pass signals across
-a tiny gap called the synapse. When an electrical impulse reaches the end of a
-neuron, it releases neurotransmitter molecules that drift across the synapse and
-bind to receptors on the next cell.
-
-The popular idea of a single "feel-good chemical" is misleading. What matters is
-the **balance** between excitatory signaling (glutamate), inhibitory signaling
-(GABA), and modulatory systems (dopamine, serotonin, acetylcholine). Receptor
-sensitivity, reuptake speed, and enzyme activity shape the response as much as
-the raw amount of any neurotransmitter.
-
-This systems view is the right lens for thinking about herbs and compounds: most
-do not "boost" one chemical so much as nudge a pathway, and individual response
-varies widely.
