--- content/education/how-stress-affects-the-brain.md
@@ -1,48 +1,0 @@
----
-slug: how-stress-affects-the-brain
-title: How Stress Affects the Brain
-summary: How acute and chronic stress reshape attention, memory, and emotional regulation through cortisol and the HPA axis — and why recovery, not just suppression, is the goal.
-category: Stress Physiology
-goals:
-  - stress
-  - sleep
-  - recovery
-  - focus
-pathways:
-  - hpa-axis
-  - cortisol
-  - glutamate
-evidenceGrade: Moderate
-safety: Educational overview; persistent stress symptoms warrant professional care.
-keywords:
-  - stress
-  - cortisol
-  - amygdala
-  - prefrontal cortex
-  - burnout
-  - allostatic load
-tags:
-  - stress
-  - recovery
-relatedHerbs:
-  - ashwagandha
-  - rhodiola
-relatedCompounds:
-  - magnesium-glycinate
-  - l-theanine
-readingTime: 9 min read
-updated: 2026-06-08
----
-
-Short bursts of stress sharpen attention and memory. The problem is **chronic**
-activation. Sustained cortisol exposure biases the brain toward the amygdala
-(threat detection) and away from the prefrontal cortex (planning and
-self-regulation), which is why prolonged stress feels like brain fog, reactivity,
-and poor sleep.
-
-Over time this "allostatic load" can blunt recovery systems. The useful reframe is
-that the target is not zero stress but **better recovery** — sleep continuity,
-downshifting the nervous system, and reducing total load.
-
-Supplements discussed in this context are best understood as small supports inside
-that larger recovery picture, not replacements for it.
