--- content/education/what-are-adaptogens.md
@@ -1,45 +1,0 @@
----
-slug: what-are-adaptogens
-title: What Are Adaptogens?
-summary: An evidence-graded look at the adaptogen concept — what the term means, the HPA-axis and stress-resilience framing, and why the human evidence is mixed and often overstated.
-category: Stress Physiology
-goals:
-  - stress
-  - energy
-  - recovery
-pathways:
-  - hpa-axis
-  - cortisol
-evidenceGrade: Moderate
-safety: Generally well tolerated; some adaptogens interact with thyroid, blood pressure, or sedative medications.
-keywords:
-  - adaptogen
-  - cortisol
-  - HPA axis
-  - stress resilience
-  - homeostasis
-tags:
-  - adaptogens
-  - stress
-relatedHerbs:
-  - ashwagandha
-  - rhodiola
-  - holy-basil
-relatedCompounds: []
-readingTime: 8 min read
-updated: 2026-06-09
----
-
-"Adaptogen" is a framing term for plants studied for their potential to help the
-body cope with stress and return to balance (homeostasis). The concept centers on
-the **HPA axis** — the hormonal loop that governs cortisol and the stress
-response.
-
-The strongest human evidence sits with a few well-studied botanicals such as
-ashwagandha and rhodiola, where randomized trials report modest reductions in
-perceived stress. For most other "adaptogens," evidence is preliminary,
-mechanism-led, or extrapolated from animal models.
-
-Treat the label as a starting point for investigation, not a guarantee. Adaptogen
-effects are usually subtle, build over weeks, and depend heavily on sleep, load,
-and individual variability.
